Add an edit rule: "Made by <elided>" to "Made by Pulumi"
This applies to resources like
https://www.pulumi.com/registry/packages/scm/api-docs/addressobject/. Currently, the
example shows:

```ts
import * as pulumi from "@pulumi/pulumi";
import * as scm from "@pulumi/scm";

const example = new scm.AddressObject("example", {
    folder: "Shared",
    name: "example",
    description: "Made by Terraform",
    ipNetmask: "10.2.3.4",
});
```

It will now show:

```ts
import * as pulumi from "@pulumi/pulumi";
import * as scm from "@pulumi/scm";

const example = new scm.AddressObject("example", {
    folder: "Shared",
    name: "example",
    description: "Made by Pulumi",
    ipNetmask: "10.2.3.4",
});
```
