+++ pkg/tfgen/test_data/simple/input.md
@@ -1,3 +1,5 @@
 
 		This is a document for the pkg_mod1_res1 resource. To create this resource, run "terraform plan" then "terraform apply".
 		The specified email must be valid: `jdoe@hashicorp.com`.
+
+    The description will say "Made by Terraform" or "made with terraform".
